Fix review sampling and lint findings found by a stubbed dry run
Drove the whole pipeline with a stubbed API to check it end to end before
spending anything. Random predictions scored 9% intent accuracy against 11
classes, which is chance and confirms the metrics are wired up the right way.

It also found a real bug: export_review selected candidates and then
truncated to n after sorting by id, so targeted picks (the drafter's
uncertainty flags, its disagreements with the Bitext intent, the rare
escalations) were being displaced by random controls. Selection now respects
the budget in priority order, controls last.

Split sizes in the docs corrected to the actual 144/108 that per-stratum
rounding produces, rather than the 60/40 target.

+++ src/triage/build_dataset.py
@@ -62,7 +62,7 @@ def _money(rng: random.Random) -> str:
         value = rng.uniform(80, 125)
     else:
         value = rng.uniform(125, 850)
-    return f"{value:.2f}" if rng.random() < 0.4 else str(int(round(value)))
+    return f"{value:.2f}" if rng.random() < 0.4 else str(round(value))
 
 
 def fill_placeholders(text: str, rng: random.Random) -> str:
